Ajuste de propriedade

+++ APICatalogo/Models/Produto.cs
@@ -9,17 +9,22 @@ namespace APICatalogo.Models
     {
         [Key]
         public int ProdutoId { get; set; }
+
         [Required]
         [MaxLength(80)]
         public string Nome { get; set; }
+
         [Required]
         [MaxLength(300)]
         public string Descricao { get; set; }
+
         [Required]
-        public decimal Preco { get; set; }
+        public double Preco { get; set; }
+
         [Required]
         [MaxLength(300)]
         public string ImgURL { get; set; }
+
         public float Estoque { get; set; }
         public DateTime DataCadastro { get; set; }
